Removed extra conditional in calculate_duration; enhanced docs
- The conditional that handles the [H, M, S] case in the code
is just a special case of the other code that handles the
other 6 cases, so I removed it for clarity and brevity.
- I clarified the documentation in calculate_duration to make
it clear that at least one of 'H', 'M', or 'S' is expected to be
present in the passed in duration string
- Reformatted docstrings throughout to make it more consistent
with the rest of the code.